% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{addEdge}
\alias{addEdge}
\title{Adding edge to graph}
\usage{
addEdge(graph, edge)
}
\arguments{
\item{graph}{a dataframe of class graph}

\item{edge}{a vector of class edge}
}
\value{
the graph with the additional edge "edge"
}
\description{
Adding edge to a graph
}
\examples{
myGraph <- graph()
myGraph <- addEdge(myGraph, edge(0, 1, "up", 10))
myGraph <- addEdge(myGraph, edge(1, 0, "down", 0))
}
