% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{addStartEnd}
\alias{addStartEnd}
\title{Constraint starting and ending states to a graph}
\usage{
addStartEnd(graph, start = -1, end = -1)
}
\arguments{
\item{graph}{a dataframe of class graph}

\item{start}{a nonnegative integer. The first state contrained in the changepoint inference}

\item{end}{a nonnegative integer. The end state contrained in the changepoint inference}
}
\value{
the graph with these new constraints
}
\description{
Adding constraints on the starting and ending states to a graph
}
\examples{
myGraph <- graph()
myGraph <- addEdge(myGraph, edge(0, 1, "up", 10))
myGraph <- addEdge(myGraph, edge(1, 0, "down", 0))
myGraph <- addStartEnd(myGraph, 0, 0)
}
