\name{ECG}
\alias{ECG}
\docType{data}
\title{
  Electrocardiogram data
}
\description{
  The goal in these data is to detect the QRS complex, as does the
  Pan-Tompkins algorithm.
}
\usage{data("ECG")}
\format{
  A named list of two data.tables: data has 1001 rows and 2 columns,
  PanTompkins has 24 rows and 3 columns.
}
\references{
  \url{https://www.robots.ox.ac.uk/~gari/teaching/cdt/A3/readings/ECG/Pan+Tompkins.pdf}
}
\keyword{datasets}
