% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{Node}
\alias{Node}
\title{Node Values}
\usage{
Node(state = NULL, min = -Inf, max = Inf)
}
\arguments{
\item{state}{a string defining the state to constrain}

\item{min}{minimal value for the inferred parameter}

\item{max}{maximal value for the inferred parameter}
}
\value{
a dataframe with 10 variables with only \code{state1}, \code{min} and \code{max} defined (not \code{NA}).
}
\description{
Constrain the range of values to consider at a node
}
\examples{
Node(state = "s0", min = 0, max = 2)

Node(state = 0, min = -1, max = 1)

Node(state = "positive", min = 0)

Node(state = "mu0", min = 0.5, max = 0.5)
}
