% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGenerator.R
\name{sdDiff}
\alias{sdDiff}
\title{sdDiff}
\usage{
sdDiff(x, method = "HALL")
}
\arguments{
\item{x}{vector of datapoint}

\item{method}{Three available methods: "HALL", "MAD" and "SD"}
}
\value{
a value equal to the estimated standard deviation
}
\description{
Estimation of the standard deviation
}
\examples{
data <- dataGenerator(100, c(0.3, 0.6, 1), c(1, 2, 3), 2)
sdDiff(data)

}
