% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.gfpop}
\alias{summary.gfpop}
\title{summary.gfpop}
\usage{
\method{summary}{gfpop}(object, ...)
}
\arguments{
\item{object}{a gfpop class object}

\item{...}{Other parameters}
}
\value{
summary of gfpop object
}
\description{
Returns all information contains in gfpop object
}
\examples{
n <- 1000 #data length
data <- dataGenerator(n, c(0.3, 0.4, 0.7, 0.95, 1), c(1, 3, 1, -1, 4), "mean", sigma = 3)
myGraph <- graph(type = "relevant", gap = 0.5, penalty = 2 * sdDiff(data) ^ 2 * log(n))
g <- gfpop(data, myGraph, type = "mean")
summary(g)
}
