% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{edge}
\alias{edge}
\title{Edge generation}
\usage{
edge(state1, state2, type = "null", penalty = 0, decay = 1,
  gap = 0, oneValue = NULL, K = Inf, a = NA, min = -Inf,
  max = Inf)
}
\arguments{
\item{state1}{a nonnegative integer defining the starting state of the edge}

\item{state2}{a nonnegative integer defining the ending state of the edge}

\item{type}{a string equal to "null", "std", "up", "down", "absInf" or "absSup"}

\item{penalty}{a nonnegative number. The penality associated to this state transition}

\item{decay}{a nonnegative number to give the strength of the exponential decay into the segment}

\item{gap}{a nonnegative number to constraint the size of the gap in the change of state}

\item{oneValue}{the value of the oneValue edge}

\item{K}{a positive number. Threshold for the Biweight robust loss}

\item{a}{a positive number. Slope for the Huber robust loss}

\item{min}{minimal value for the infered parameter}

\item{max}{maximal value for the infered parameter}
}
\value{
a dataframe with five components equal to the five parameters
}
\description{
Edge creation for graph
}
\examples{
edge("D", "U", "up", penalty = 10, gap = 1)
}
