% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{edge}
\alias{edge}
\title{Edge generation}
\usage{
edge(state1, state2, type = "null", penalty = 0, parameter = 1)
}
\arguments{
\item{state1}{a nonnegative integer defining the starting state of the edge}

\item{state2}{a nonnegative integer defining the ending state of the edge}

\item{type}{a string equal to "null", "std", "up", "down", "absInf" or "absSup"}

\item{penalty}{a nonnegative number. The penality associated to this state transition}

\item{parameter}{a nonnegative number to constraint the size of the gap in the change of state}
}
\value{
a dataframe with five components equal to the five parameters
}
\description{
Edge creation for graph
}
\examples{
edge(0, 1, "up", 10, 1)
}
