% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfpop.R
\name{gfpop}
\alias{gfpop}
\title{Graph-constrained functional pruning optimal partitioning}
\usage{
gfpop(data = c(0), mygraph, type = "mean", weights = c(0))
}
\arguments{
\item{mygraph}{dataframe of class graph to constraint the changepoint dynamic programming algorithm}

\item{type}{a string defining the type of cost to use. "gauss", "poisson" or "binomial"}

\item{vectData}{vector of data to segment}

\item{vectWeight}{vector of weights (positive numbers) same size as vectData}

\item{K}{a positive number. Threshold for the Biweight robust loss}

\item{a}{a positive number. Slope for the Huber robust loss}

\item{min}{minimal bound for the infered means}

\item{max}{maximal bound for the infered means}
}
\value{
a gfpop object = (changepoints, states, forced, means).
'changepoints' is the vector of changepoints (we give the last element of each segment).
'states' is the vector giving the state of each segment
'forced' is the vector specifying whether the constraints of the graph are active (=1) or not (=0)
'means' is the vector of successive means of each segment
'cost' is a number equal to the global cost of the graph-constrained segmentation
}
\description{
Graph-contstrained functional pruning optimal partitionning
}
