% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfpop.R
\name{gfpop}
\alias{gfpop}
\title{Graph-Constrained Functional Pruning Optimal Partitioning}
\usage{
gfpop(data = c(0), mygraph, type = "mean", weights = c(0))
}
\arguments{
\item{data}{vector of data to segment}

\item{mygraph}{dataframe of class "graph" to constrain the changepoint inference}

\item{type}{a string defining the cost model to use: "mean", "variance", "poisson", "exp", "negbin"}

\item{weights}{vector of weights (positive numbers), same size as data}
}
\value{
a gfpop object = (changepoints, states, forced, parameters, globalCost)
'changepoints' is the vector of changepoints (we give the last element of each segment)
'states' is the vector giving the state of each segment
'forced' is the vector specifying whether the constraints of the graph are active (=1) or not (=0)
'parameters' is the vector of successive parameters of each segment
'globalCost' is a number equal to the global cost of the graph-constrained changepoint optimization problem
}
\description{
Functional pruning optimal partitioning with a graph structure to take into account constraints on consecutive segment parameters. The user has to specify the graph he wants to use (see the graph function) and a type of cost funcion. This is the main function of the gfpop package.
}
