% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph}
\alias{graph}
\title{Graph generation}
\usage{
graph(penalty = 0, type = "empty")
}
\arguments{
\item{penalty}{a nonnegative number equals to the common penalty to use for all edges}

\item{type}{a string equal to "std", "isotonic", "updown", "infsup". to build a predefined classic graph}
}
\value{
a dataframe with edges in rows (columns are named "state1", "state2", "type", "penalty", "parameter") with additional "graph" class.
}
\description{
Graph creation
}
\examples{
myGraph <- graph(penalty = 10, "updown")
myEmptyGraph <- graph()
}
