% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph}
\alias{graph}
\title{Graph generation}
\usage{
graph(..., type = "empty", penalty = 0, decay = 1, gap = 0)
}
\arguments{
\item{...}{This is a list of edges definied by functions "Edge", "StartEnd" and "Node"}

\item{type}{a string equal to "std", "isotonic", "updown", "infsup". to build a predefined classic graph}

\item{penalty}{a nonnegative number equals to the common penalty to use for all edges}

\item{decay}{a nonnegative number to give the strength of the exponential decay into the segment}

\item{gap}{a nonnegative number to constrain the size of the gap in the change of state}
}
\value{
a dataframe with 9 variables (columns are named "state1", "state2", "type", "penalty", "parameter", "K", "a", "min", "max") with additional "graph" class.
}
\description{
Graph creation using component functions "Edge", "StartEnd" and "Node"
}
\examples{
UpDownGraph <- graph(type = "updown", penalty = 10, gap = 1.3)
MyGraph <- graph(Edge(0,0), Edge(1,1), Edge(0,1,"up",10,gap=0.5),
Edge(1,0,"down"), StartEnd(0,0), Node(0,0,1), Node(1,0,1))
}
