% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph}
\alias{graph}
\title{Graph generation}
\usage{
graph(..., penalty = 0, type = "empty")
}
\arguments{
\item{...}{This is a list of edges definied by functions edge and StartEnd}

\item{penalty}{a nonnegative number equals to the common penalty to use for all edges}

\item{type}{a string equal to "std", "isotonic", "updown", "infsup". to build a predefined classic graph}
}
\value{
a dataframe with edges in rows (columns are named "state1", "state2", "type", "penalty", "parameter") with additional "graph" class.
}
\description{
Graph creation
}
\examples{
UpDownGraph <- graph(penalty = 10, type = "updown")
MyGraph <- graph(edge(0,0), edge(1,1), edge(0,1,"up",10,gap=0.5), edge(1,0,"down"), StartEnd(0,0))
}
