% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{Node}
\alias{Node}
\title{Node Values}
\usage{
Node(state = NULL, min = -Inf, max = Inf)
}
\arguments{
\item{state}{a string defining the state to constrain}

\item{min}{minimal value for the infered parameter}

\item{max}{maximal value for the infered parameter}
}
\value{
a dataframe with 9 variables with only `state1`, `min` and `max` defined.
}
\description{
Constraint the rabge of values to consider for a node
}
\examples{
Node(state = "S", min = 0, max = 2)
}
