% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paperGraphs.R
\name{paperGraph}
\alias{paperGraph}
\title{Graphs of our paper in jss}
\usage{
paperGraph(nb)
}
\arguments{
\item{nb}{the number of the Figure in paper}
}
\value{
a graph
}
\description{
Graphs of our paper in jss
}
