% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paperGraphs.R
\name{paperGraph}
\alias{paperGraph}
\title{Graphs of our paper in jss}
\usage{
paperGraph(nb, penalty = 0, decay = 1, gap = 0, oneValue = 0,
  K = Inf, a = 0)
}
\arguments{
\item{nb}{the number of the Figure in paper}

\item{penalty}{the penalty to use for change-point}

\item{decay}{a nonnegative number to give the strength of the exponential decay into the segment}

\item{gap}{a nonnegative number to constrain the size of the gap in the change of state}

\item{oneValue}{the value for parameter when we consider the collective anomalies problem}

\item{K}{a positive number. Threshold for the Biweight robust loss}

\item{a}{a positive number. Slope for the Huber robust loss}
}
\value{
a dataframe with 9 variables (columns are named "state1", "state2", "type", "parameter", "penalty", "K", "a", "min", "max") with additional "graph" class.
}
\description{
Graphs of our paper in jss
}
