% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.gfpop}
\alias{plot.gfpop}
\title{plot.gfpop}
\usage{
\method{plot}{gfpop}(x, ..., data)
}
\arguments{
\item{x}{a gfpop class object}

\item{...}{Other parameters}

\item{data}{the data from which we get the gfpop result}
}
\value{
plot data and the inferred gfpop segments
}
\description{
Plot the result of the gfpop function with the data-points
}
