% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{StartEnd}
\alias{StartEnd}
\title{Start and End nodes for the graph}
\usage{
StartEnd(start = NULL, end = NULL)
}
\arguments{
\item{start}{a vector of states. The beginning vertices in the changepoint inference}

\item{end}{a vector of states. The ending vertices in the changepoint inference}
}
\value{
dataframe with 9 variables with only `state1` and `type` = start or end defined.
}
\description{
Adding constraints on the beginning and ending states of a graph
}
\examples{
StartEnd(start = "A", end = c("A","B"))
}
