% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{StartEnd}
\alias{StartEnd}
\title{Constraint starting and ending states to a graph}
\usage{
StartEnd(start = -1, end = -1)
}
\arguments{
\item{start}{a nonnegative integer. The first vertex in the changepoint inference}

\item{end}{a nonnegative integer. The end vertex in the changepoint inference}
}
\value{
a dataframe with five components (as for edge) with only state1 and type = start or end defined.
}
\description{
Adding constraints on the starting and ending states to a graph
}
\examples{
StartEnd(start = 0, end = 1)
}
